Don't set m_bssl to true until we have actually set SSL up
If for some reason we return out of Csock::SSLServerSetup() before SSL
is setup, for example because we don't have access to the PEM file. Then
we should leave m_bssl to false else if the socket sends us data we will
try to use SSL_read() which will cause a segmentation fault.