Fix use of a feature deprecated in OpenSSL 1.0
If OpenSSL 1.1 is configured in such a way that features deprecated in 1.0.0 are not built, then compile errors would result here.
If it is also configured in such a way that features deprecated in 1.1.0 are not built, then compile errors would result from the replacement API (which was of course deprecated in 1.1).