Fix use of features deprecated in OpenSSL 1.1
If OpenSSL 1.1 is configured in such a way that features deprecated in 1.1.0 are not built, then compile errors would result here.

If I'm reading the previous algorithm correctly then this changes the meaning of `m_iMethod` in `SSLClientSetup` and `SetupServerCTX`:
- By default, it uses the version-flexible SSLv23_*_method/TLS_*_method and trusts the SSL library to do the right thing. [unchanged]
- For non-default `m_iMethod` where the SSL library supports that version, then it uses that version only. [unchanged]
- For non-default `m_iMethod` where the SSL library does not support that version, now it always uses the version-flexible SSLv23_*_method/TLS_*_method and trusts the SSL library to do the right thing. [changed?]

It also looks like there's a minor bug in the old algorithm in `SetupServerCTX` where asking for TLS1 with an SSL library not supporting TLS1.1 would bypass the TLS1 case and fall through to the default case (`case TLS1:` is inside the TLS1_1_VERSION guard).