Add regression test for recently added OpenSSL 1.1 support
Requires `openssl` and `sed` binaries to be on the path, which should
be fine considering an SSL library is needed for the test and sed is
already used in the makefile.

The openssl binary should support the same protocol versions as the SSL
library used to build Csocket and this test, otherwise the protocol
version checks will fail.