Simplify "is this utf8?"-check
Currently the code tries to convert from UTF-8 to UTF-8 in strict mode
to check if a given string is valid. As suggested in [1], just checking
if a given string is valid UTF-8 can be done in a simpler way via
u_strFromUTF8, a function that converts from UTF-8 to UTF-16, but
calling it in such a way that it does not actually produce any output.

[1]: https://sourceforge.net/p/icu/mailman/message/32802299/
Signed-off-by: Uli Schlachter <psychon@znc.in>